xfconf: fix 'Failed to get connection to xfconfd' during do_rootfs
Packages such as gdk-pixbuf can result in the following error during
the rootfs construction:
  CRITICAL **: 16:51:57.223: Failed to get connection to xfconfd: Cannot autolaunch D-Bus without X11 $DISPLAY

This is hit during the postinst-intercepts/update_pixbuf_cache.

Based on discussions upstream:

https://bugzilla.xfce.org/show_bug.cgi?id=14789

this appears to be an issue for xfconf and has an upstream update:

https://git.xfce.org/xfce/xfconf/commit?id=6f11cc65a2213f424728b5a0d25b1b63c19e773e

Drawing from this change we can go ahead and use PACKAGECONFIG to
emulate the upstream fix, defaulting to having the gsettings-backend
disabled.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>